requirements.txt
tqdm==4.24.0
visdom==0.1.8.5
pandas==0.22.0
matplotlib==2.1.2
torch==0.4.1
numpy==1.14.2
Pillow==5.3.0
imageio==2.4.1
torchvision==0.2.1
